feat : front)  지도 컴포넌트화
